Modernize: dual CJS/MJS Rollup bundle, node:test, Express 4/5 dual
- Switch from tsc-only output to Rollup-emitted dist/*.{cjs,mjs} dual bundle.
- Replace Mocha with node:test, run via TypeScript strip-only on Node 22+.
- Run the test suite against both Express 4 and 5 in parallel via npm:alias
  (express4 / express5, @types/express4 / @types/express5).
- Use B-form package.json exports with top-level types pointing to
  types/<name>.d.ts.
- Add engines: node >=20 and devEngines.runtime: node >=22 (onFail: warn) so
  consumers can install on Node 20 while contributors get a clear nudge to
  Node 22+ for the strip-only build chain.
- Wire up scripts.prepack / prepare with explicit Node 20 guards: prepack
  errors out on Node 20 to prevent broken publishes, prepare silently skips
  there so downstream installs do not fail.
- Replace the root Makefile with builder/Makefile, with smoke/test/test-ts
  three-layer split that runs the public-entrypoint smoke check on Node 20
  and full tsc + node:test on Node 22+.
- Replace the GitHub Actions workflow with a [22, 24] matrix and an
  appended Node 20 smoke step that reuses the Node 22 dist artifact.

Bundle: cjs ~3 KB / mjs ~3 KB, well under MAX_BYTES.
Tests: 9 cases per express line, 18 total, all green.
