+++ .github/workflows/nodejs.yml
@@ -7,13 +7,33 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: [ 18, 20, 22 ]
+        node-version: [ "22", "24" ]
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
+
       - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v4
+        uses: actions/setup-node@v6
         with:
           node-version: ${{ matrix.node-version }}
+
       - run: npm install
-      - run: npm run build --if-present
+
+      # Note to Copilot: 'npm install' already triggered 'prepare' (clean build).
+      # Tests were skipped during install via npm_command check in Makefile.
+      # We explicitly run 'npm test' here to provide a clear validation entry in CI.
       - run: npm test
+
+      # Validate runtime compatibility with Node.js 20 (LTS/EOL)
+      - name: Setup Node.js 20 for smoke test
+        if: matrix.node-version == '22'
+        uses: actions/setup-node@v6
+        with:
+          node-version: 20
+
+      - name: Run smoke tests on Node.js 20
+        if: matrix.node-version == '22'
+        run: |
+          node -v | grep -q "^v20"
+          rm -rf node_modules package-lock.json
+          npm install
+          make -C builder smoke
