+++ builder/Makefile
@@ -1,0 +1,67 @@
+SRC := $(wildcard ../lib/*.ts)
+TYPES := $(wildcard ../types/*.d.ts)
+BUILDER_SRC := $(wildcard ./*.ts)
+
+MJS_DST := ../dist/async-request-handler.mjs
+CJS_DST := ../dist/async-request-handler.cjs
+
+MJS_CONFIG := ./rollup-mjs.config.ts
+CJS_CONFIG := ./rollup-cjs.config.ts
+
+# Bundle is small (a thin wrapper over express-intercept + sed-lite, both
+# kept external). Anything noticeably larger means an accidental bundling
+# of a Node built-in or a transitive dep — fail the build so it cannot
+# slip in.
+MAX_BYTES := 5000
+
+# Every named export the package documents must survive the round trip
+# through both CJS and ESM bundles. List them once and assert in both
+# smoke checks below.
+NAMED_EXPORTS := ASYNC CATCH IF
+
+ALL := $(MJS_DST) $(CJS_DST)
+
+####
+
+all: $(ALL)
+
+$(MJS_DST): $(MJS_CONFIG) $(SRC) $(BUILDER_SRC)
+	@mkdir -p $(dir $@)
+	../node_modules/.bin/rollup -c $<
+	@test "$$(wc -c < $@)" -le $(MAX_BYTES) || { echo "ERROR: $@ exceeds $(MAX_BYTES) byte cap" >&2; exit 1; }
+	@ls -l $@
+	@echo ""
+
+$(CJS_DST): $(CJS_CONFIG) $(SRC) $(BUILDER_SRC)
+	@mkdir -p $(dir $@)
+	../node_modules/.bin/rollup -c $<
+	@test "$$(wc -c < $@)" -le $(MAX_BYTES) || { echo "ERROR: $@ exceeds $(MAX_BYTES) byte cap" >&2; exit 1; }
+	@ls -l $@
+	@echo ""
+
+# Verify the bundles by importing/requiring the package itself.
+# This ensures that the public entrypoints (exports/main) are correctly configured.
+# Note: process.argv.slice(1) is used because node -e 'code' arg1 arg2
+# results in process.argv = ['node', 'arg1', 'arg2'].
+smoke: all
+	@# MJS smoke test via public entrypoint
+	cd .. && node --input-type=module -e 'const m = await import("async-request-handler"); for (const k of process.argv.slice(1)) { if (typeof m[k] !== "function") { console.error("missing MJS export:", k); process.exit(1); } console.log("MJS export OK:", k); }' $(NAMED_EXPORTS)
+	@# CJS smoke test via public entrypoint
+	cd .. && node --input-type=commonjs -e 'const m = require("async-request-handler"); for (const k of process.argv.slice(1)) { if (typeof m[k] !== "function") { console.error("missing CJS export:", k); process.exit(1); } console.log("CJS export OK:", k); }' $(NAMED_EXPORTS)
+
+####
+
+# Build, run smoke tests, and conditionally run full tests (tsc and node:test).
+test: all smoke $(TYPES)
+	@node -v | grep -qv "^v20" || echo "Node 20 detected: skipping full tests."
+	@ [ "$$npm_command" != "install" ] || echo "npm install detected: skipping full tests."
+	@node -v | grep -q "^v20" || [ "$$npm_command" = "install" ] || $(MAKE) test-ts
+
+test-ts:
+	../node_modules/.bin/tsc --noEmit
+	cd .. && node --test test/*.test.ts
+
+clean:
+	/bin/rm -fr $(ALL)
+
+.PHONY: all clean test smoke test-ts
