+++ test/basic.express4.test.ts
@@ -1,0 +1,10 @@
+// Test entry for the Express 4 line.
+
+import {describe} from "node:test";
+import express from "express4";
+
+import {runBasicTests} from "./lib/basic.ts";
+
+describe("basic.express4.test.ts", () => {
+    runBasicTests(express);
+});
