--- tsconfig-esm.json
@@ -1,20 +1,0 @@
-{
-  "include": [
-    "./lib",
-    "./test"
-  ],
-  "compilerOptions": {
-    "module": "ES2022",
-    "moduleResolution": "Bundler",
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": false,
-    "outDir": "build/",
-    "removeComments": false,
-    "skipLibCheck": true,
-    "strictNullChecks": false,
-    "target": "ES2022"
-  }
-}
